SX Design: switch `LinkButton` and `ButtonLink` implementations
Closes: https://github.com/adeira/universe/issues/2691
