+++ src/abacus-backoffice/src/LayoutHeadingLink.js
@@ -2,7 +2,7 @@
 
 import { useRouter } from 'next/router';
 import React, { type Node } from 'react';
-import { ButtonLink_DEPRECATED as ButtonLink } from '@adeira/sx-design';
+import { LinkButton } from '@adeira/sx-design';
 import NextLink from 'next/link';
 
 type Props = {
@@ -15,7 +15,7 @@ export default function LayoutHeadingLink(props: Props): Node {
 
   return (
     <NextLink href={props.href} locale={router.locale} passHref={true}>
-      <ButtonLink href={props.href}>{props.children}</ButtonLink>
+      <LinkButton href={props.href}>{props.children}</LinkButton>
     </NextLink>
   );
 }
