+++ src/sx-design/src/Button/__tests__/ButtonLink.test.js
@@ -3,46 +3,30 @@
  * @jest-environment jsdom
  */
 
-import { fireEvent } from '@testing-library/react';
 import React from 'react';
-import '@adeira/sx-jest-snapshot-serializer';
 import fbt from 'fbt';
 
-import ButtonLink from '../ButtonLink_DEPRECATED';
-import { initFbt, render } from '../../test-utils';
+import ButtonLink from '../ButtonLink';
+import { initFbt, render, fireEvent } from '../../test-utils';
 
 beforeEach(() => {
   initFbt();
-});
-
-it('forwards React refs as expected', () => {
-  const ref = React.createRef();
-  const { container } = render(
-    <ButtonLink ref={ref} href="https://localhost">
-      <fbt desc="button link title" doNotExtract={true}>
-        button link
-      </fbt>
-    </ButtonLink>,
-  );
-
-  expect(ref.current).toBe(container.firstChild);
-  expect(ref.current?.nodeName).toBe('A');
 });
 
 it('calls onClick event', () => {
   const onClickFn = jest.fn();
 
   const { getByText } = render(
-    <ButtonLink href="https://localhost" onClick={onClickFn}>
-      <fbt desc="button link title" doNotExtract={true}>
-        button link with onClick callback
+    <ButtonLink onClick={onClickFn}>
+      <fbt desc="link title" doNotExtract={true}>
+        link button with onClick callback
       </fbt>
     </ButtonLink>,
   );
 
-  expect(getByText('button link with onClick callback')).toBeDefined();
+  expect(getByText('link button with onClick callback')).toBeDefined();
   expect(onClickFn).not.toHaveBeenCalled();
 
-  fireEvent.click(getByText('button link with onClick callback'));
+  fireEvent.click(getByText('link button with onClick callback'));
   expect(onClickFn).toHaveBeenCalledTimes(1);
 });
