+++ src/sx-design/src/Link/LinkButton.js
@@ -3,8 +3,8 @@
 import * as React from 'react';
 import sx from '@adeira/sx';
 
-import Link from '../Link/Link';
-import sharedButtonStyles from './styles';
+import Link from './Link';
+import sharedButtonStyles from '../Button/styles';
 
 type Props = {
   +'href': string,
@@ -24,7 +24,7 @@ type Props = {
  * Optionally, you can use [React refs](https://reactjs.org/docs/refs-and-the-dom.html) and it will
  * be forwarded to the HTML `<a />` element as expected.
  */
-export default (React.forwardRef(function ButtonLink_DEPRECATED(props, ref): React.Node {
+export default (React.forwardRef(function LinkButton(props, ref): React.Node {
   return (
     <Link
       ref={ref}
