+++ src/sx-design/src/Link/__tests__/LinkButton.test.js
@@ -3,30 +3,46 @@
  * @jest-environment jsdom
  */
 
+import { fireEvent } from '@testing-library/react';
 import React from 'react';
+import '@adeira/sx-jest-snapshot-serializer';
 import fbt from 'fbt';
 
-import LinkButton_DEPRECATED from '../LinkButton_DEPRECATED';
-import { initFbt, render, fireEvent } from '../../test-utils';
+import LinkButton from '../LinkButton';
+import { initFbt, render } from '../../test-utils';
 
 beforeEach(() => {
   initFbt();
+});
+
+it('forwards React refs as expected', () => {
+  const ref = React.createRef();
+  const { container } = render(
+    <LinkButton ref={ref} href="https://localhost">
+      <fbt desc="button link title" doNotExtract={true}>
+        button link
+      </fbt>
+    </LinkButton>,
+  );
+
+  expect(ref.current).toBe(container.firstChild);
+  expect(ref.current?.nodeName).toBe('A');
 });
 
 it('calls onClick event', () => {
   const onClickFn = jest.fn();
 
   const { getByText } = render(
-    <LinkButton_DEPRECATED onClick={onClickFn}>
-      <fbt desc="link title" doNotExtract={true}>
-        link button with onClick callback
+    <LinkButton href="https://localhost" onClick={onClickFn}>
+      <fbt desc="button link title" doNotExtract={true}>
+        button link with onClick callback
       </fbt>
-    </LinkButton_DEPRECATED>,
+    </LinkButton>,
   );
 
-  expect(getByText('link button with onClick callback')).toBeDefined();
+  expect(getByText('button link with onClick callback')).toBeDefined();
   expect(onClickFn).not.toHaveBeenCalled();
 
-  fireEvent.click(getByText('link button with onClick callback'));
+  fireEvent.click(getByText('button link with onClick callback'));
   expect(onClickFn).toHaveBeenCalledTimes(1);
 });
